adds /dataModels and Gene Model
* MockBackend: moved to separate angular module
* GenesResourceFactory: re-factored  factory to use $http instead of $resource and use GeneModel
* Gene: master Gene class
*GeneModel: Gene class data model to transform responses
