+++ .babelrc
@@ -1,6 +1,7 @@
 {
   "presets": ["es2015"],
   "plugins": [
-    "add-module-exports"
+    "add-module-exports",
+    "transform-class-properties"
   ]
 }
