--- app/js/.factories/example.js
@@ -1,13 +1,0 @@
-function ExampleFactory($http) {
-  'ngInject';
-
-  const Factory = {};
-
-  return Factory;
-
-}
-
-export default {
-  name: 'ExampleFactory',
-  fn: ExampleFactory
-};
