--- app/js/MockBackend/mockData/genes.js
@@ -1,303 +1,0 @@
-module.exports = {
-  'total': 37,
-  'max_score': 17.74919,
-  'took': 15,
-  'hits': [
-    {
-      '_id': '4331',
-      '_score': 17.74919,
-      'entrezgene': 4331,
-      'name': 'MNAT1, CDK activating kinase assembly factor',
-      'symbol': 'MNAT1',
-      'taxid': 9606
-    },
-    {
-      '_id': '388324',
-      '_score': 15.771275,
-      'entrezgene': 388324,
-      'name': 'inhibitor of CDK, cyclin A1 interacting protein 1',
-      'symbol': 'INCA1',
-      'taxid': 9606
-    },
-    {
-      '_id': '1030',
-      '_score': 0.9556542,
-      'entrezgene': 1030,
-      'name': 'cyclin dependent kinase inhibitor 2B',
-      'symbol': 'CDKN2B',
-      'taxid': 9606
-    },
-    {
-      '_id': '143384',
-      '_score': 0.94585234,
-      'entrezgene': 143384,
-      'name': 'CDK2 associated cullin domain 1',
-      'symbol': 'CACUL1',
-      'taxid': 9606
-    },
-    {
-      '_id': '1022',
-      '_score': 0.8689418,
-      'entrezgene': 1022,
-      'name': 'cyclin dependent kinase 7',
-      'symbol': 'CDK7',
-      'taxid': 9606
-    },
-    {
-      '_id': '56647',
-      '_score': 0.84865373,
-      'entrezgene': 56647,
-      'name': 'BRCA2 and CDKN1A interacting protein',
-      'symbol': 'BCCIP',
-      'taxid': 9606
-    },
-    {
-      '_id': '23552',
-      '_score': 0.66218877,
-      'entrezgene': 23552,
-      'name': 'cyclin dependent kinase 20',
-      'symbol': 'CDK20',
-      'taxid': 9606
-    },
-    {
-      '_id': '1026',
-      '_score': 0.6565289,
-      'entrezgene': 1026,
-      'name': 'cyclin dependent kinase inhibitor 1A',
-      'symbol': 'CDKN1A',
-      'taxid': 9606
-    },
-    {
-      '_id': '902',
-      '_score': 0.6432643,
-      'entrezgene': 902,
-      'name': 'cyclin H',
-      'symbol': 'CCNH',
-      'taxid': 9606
-    },
-    {
-      '_id': '1032',
-      '_score': 0.63649035,
-      'entrezgene': 1032,
-      'name': 'cyclin dependent kinase inhibitor 2D',
-      'symbol': 'CDKN2D',
-      'taxid': 9606
-    },
-    {
-      '_id': '1033',
-      '_score': 0.6000888,
-      'entrezgene': 1033,
-      'name': 'cyclin dependent kinase inhibitor 3',
-      'symbol': 'CDKN3',
-      'taxid': 9606
-    },
-    {
-      '_id': '8812',
-      '_score': 0.5351293,
-      'entrezgene': 8812,
-      'name': 'cyclin K',
-      'symbol': 'CCNK',
-      'taxid': 9606
-    },
-    {
-      '_id': '1021',
-      '_score': 0.49154764,
-      'entrezgene': 1021,
-      'name': 'cyclin dependent kinase 6',
-      'symbol': 'CDK6',
-      'taxid': 9606
-    },
-    {
-      '_id': '9134',
-      '_score': 0.48227826,
-      'entrezgene': 9134,
-      'name': 'cyclin E2',
-      'symbol': 'CCNE2',
-      'taxid': 9606
-    },
-    {
-      '_id': '1025',
-      '_score': 0.41338134,
-      'entrezgene': 1025,
-      'name': 'cyclin dependent kinase 9',
-      'symbol': 'CDK9',
-      'taxid': 9606
-    },
-    {
-      '_id': '1027',
-      '_score': 0.4012912,
-      'entrezgene': 1027,
-      'name': 'cyclin dependent kinase inhibitor 1B',
-      'symbol': 'CDKN1B',
-      'taxid': 9606
-    },
-    {
-      '_id': '1024',
-      '_score': 0.4012912,
-      'entrezgene': 1024,
-      'name': 'cyclin dependent kinase 8',
-      'symbol': 'CDK8',
-      'taxid': 9606
-    },
-    {
-      '_id': '10983',
-      '_score': 0.37839356,
-      'entrezgene': 10983,
-      'name': 'cyclin I',
-      'symbol': 'CCNI',
-      'taxid': 9606
-    },
-    {
-      '_id': '901',
-      '_score': 0.37834096,
-      'entrezgene': 901,
-      'name': 'cyclin G2',
-      'symbol': 'CCNG2',
-      'taxid': 9606
-    },
-    {
-      '_id': '51265',
-      '_score': 0.3637087,
-      'entrezgene': 51265,
-      'name': 'cyclin dependent kinase like 3',
-      'symbol': 'CDKL3',
-      'taxid': 9606
-    },
-    {
-      '_id': '8558',
-      '_score': 0.3637087,
-      'entrezgene': 8558,
-      'name': 'cyclin dependent kinase 10',
-      'symbol': 'CDK10',
-      'taxid': 9606
-    },
-    {
-      '_id': '900',
-      '_score': 0.34290788,
-      'entrezgene': 900,
-      'name': 'cyclin G1',
-      'symbol': 'CCNG1',
-      'taxid': 9606
-    },
-    {
-      '_id': '1031',
-      '_score': 0.34102222,
-      'entrezgene': 1031,
-      'name': 'cyclin dependent kinase inhibitor 2C',
-      'symbol': 'CDKN2C',
-      'taxid': 9606
-    },
-    {
-      '_id': '8900',
-      '_score': 0.34102222,
-      'entrezgene': 8900,
-      'name': 'cyclin A1',
-      'symbol': 'CCNA1',
-      'taxid': 9606
-    },
-    {
-      '_id': '466',
-      '_score': 0.33109438,
-      'entrezgene': 466,
-      'name': 'activating transcription factor 1',
-      'symbol': 'ATF1',
-      'taxid': 9606
-    },
-    {
-      '_id': '894',
-      '_score': 0.33104834,
-      'entrezgene': 894,
-      'name': 'cyclin D2',
-      'symbol': 'CCND2',
-      'taxid': 9606
-    },
-    {
-      '_id': '1028',
-      '_score': 0.33104834,
-      'entrezgene': 1028,
-      'name': 'cyclin dependent kinase inhibitor 1C',
-      'symbol': 'CDKN1C',
-      'taxid': 9606
-    },
-    {
-      '_id': '2967',
-      '_score': 0.3000444,
-      'entrezgene': 2967,
-      'name': 'general transcription factor IIH subunit 3',
-      'symbol': 'GTF2H3',
-      'taxid': 9606
-    },
-    {
-      '_id': '898',
-      '_score': 0.3000444,
-      'entrezgene': 898,
-      'name': 'cyclin E1',
-      'symbol': 'CCNE1',
-      'taxid': 9606
-    },
-    {
-      '_id': '905',
-      '_score': 0.29230478,
-      'entrezgene': 905,
-      'name': 'cyclin T2',
-      'symbol': 'CCNT2',
-      'taxid': 9606
-    },
-    {
-      '_id': '2580',
-      '_score': 0.29230478,
-      'entrezgene': 2580,
-      'name': 'cyclin G associated kinase',
-      'symbol': 'GAK',
-      'taxid': 9606
-    },
-    {
-      '_id': '8899',
-      '_score': 0.27854162,
-      'entrezgene': 8899,
-      'name': 'pre-mRNA processing factor 4B',
-      'symbol': 'PRPF4B',
-      'taxid': 9606
-    },
-    {
-      '_id': '896',
-      '_score': 0.27854162,
-      'entrezgene': 896,
-      'name': 'cyclin D3',
-      'symbol': 'CCND3',
-      'taxid': 9606
-    },
-    {
-      '_id': '1019',
-      '_score': 0.27854162,
-      'entrezgene': 1019,
-      'name': 'cyclin dependent kinase 4',
-      'symbol': 'CDK4',
-      'taxid': 9606
-    },
-    {
-      '_id': '595',
-      '_score': 0.25718093,
-      'entrezgene': 595,
-      'name': 'cyclin D1',
-      'symbol': 'CCND1',
-      'taxid': 9606
-    },
-    {
-      '_id': '1029',
-      '_score': 0.23646308,
-      'entrezgene': 1029,
-      'name': 'cyclin dependent kinase inhibitor 2A',
-      'symbol': 'CDKN2A',
-      'taxid': 9606
-    },
-    {
-      '_id': '1017',
-      '_score': 0.18569441,
-      'entrezgene': 1017,
-      'name': 'cyclin dependent kinase 2',
-      'symbol': 'CDK2',
-      'taxid': 9606
-    }
-  ]
-};
