+++ app/js/components/app/app.component.js
@@ -1,8 +1,17 @@
 const AppComponent = {
     templateUrl: 'app/app.tpl.html',
     bindings: {},
-    controller: [()=>{
+    controller: ['$rootScope',function($rootScope){
         'ngInject';
+
+        $rootScope.queryBuilder = {
+        	query:{
+        		mutations: [],
+        		diseases: [{id: 'ADRENOCORTICAL CARCINOMA', positives: 10, negatives: 20},{id: 'ADRENOCORTICAL CARCINOMA', positives: 10, negatives: 20},{id: 'ADRENOCORTICAL CARCINOMA', positives: 10, negatives: 20},{id: 'ADRENOCORTICAL CARCINOMA', positives: 10, negatives: 20}]
+        	}
+        };
+
+
    }]
 }
