+++ app/js/components/infoBox/infoBox.component.js
@@ -6,6 +6,8 @@ const infoBoxComponent = {
 	transclude: true,    
     controller: function () {
     	'ngInject';
+        
+
     }
 }
