+++ app/js/components/queryBuilder/queryBuilder.component.js
@@ -3,10 +3,10 @@ const QueryBuilderComponent = {
     bindings: {},
     controller: ['$rootScope','_',function($rootScope, _) {
             	'ngInject';
+
             	this.diseaseList = [{id: 'ADRENOCORTICAL CARCINOMA', positives: 10, negatives: 20},{id: 'ADRENOCORTICAL CARCINOMA', positives: 10, negatives: 20},{id: 'ADRENOCORTICAL CARCINOMA', positives: 10, negatives: 20},{id: 'ADRENOCORTICAL CARCINOMA', positives: 10, negatives: 20}];
             	this.mutationList = [];
-
-
+                
                 $rootScope.$on('mutationSet:add', (e,mutation)=>{
                     this.mutationList.push(mutation);
                 });
@@ -27,7 +27,6 @@ const QueryBuilderComponent = {
                     }else{
                         this.mutationList.reverse();
                     }
-                    
                 });
