+++ package.json
@@ -46,6 +46,7 @@
     "babel-core": "^6.3.26",
     "babel-eslint": "^5.0.0",
     "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-transform-class-properties": "^6.19.0",
     "babel-preset-es2015": "^6.3.13",
     "babel-register": "^6.5.2",
     "babelify": "^7.2.0",
