+++ test/unit/components/infoBox/infoBox_spec.js
@@ -36,5 +36,4 @@ describe('UNIT::component: infoBox', () => {
  
     
 
-
 });
